+++ src/ui/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod date_selector;
 pub(crate) mod debug;
 pub(crate) mod gameday;
 pub(crate) mod help;
+pub(crate) mod input_popup;
 pub(crate) mod layout;
 pub(crate) mod linescore;
 pub(crate) mod logs;